requirements.txt
alabaster==0.7.12
amqp==2.6.1
argon2==0.1.10
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
blinker==1.4
celery==4.4.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
dnspython==1.15.0
dnspython3==1.15.0
docutils==0.16
email-validator==1.1.2
et-xmlfile==1.0.1
flake8==3.8.4
flake8-coding==1.3.2
Flask==1.1.2
Flask-Assets==2.0
Flask-Babel==2.0.0
Flask-Caching==1.9.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-oidc2==1.5.0
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fpdf2==2.4.2
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
ipython==7.31.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
kombu==4.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.2.2
openpyxl==2.6.4
packaging==20.9
parso==0.8.1
pep8==1.7.1
pexpect==4.8.0
phonenumbers==8.12.18
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.8.0
pyinotify==0.9.6
pyjwkest==1.4.2
pyopenssl==23.2.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySimpleSOAP==1.16.2
pytest==6.2.2
python-magic==0.4.22
pytz==2021.1
rcssmin==1.0.6
redis==3.5.3
requests==2.25.1
rjsmin==1.1.0
simplegeneric==0.8.1
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2
Sphinx==3.5.1
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.23
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.5
uWSGI==2.0.19.1
vine==1.3.0
wcwidth==0.2.5
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.3
